requirements.txt
Cerberus==1.3.4
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
PyMySQL==1.0.2
python-dotenv==0.20.0
SQLAlchemy==1.4.39
validators==0.20.0
